refactor: split model.ts into shared types.ts + keys.ts
